requirements.txt
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.0.1
idna==2.10
nest-asyncio==1.5.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
ujson==4.0.2
urllib3==1.26.4
